feat(init): move resolveTemplates to packages/luca-framework for importability
Move the core template resolution functions (resolveContent, resolveFilePath,
resolveTemplates, BrandingContext) from scripts/resolve-templates.ts into
packages/luca-framework/src/utils/resolve-templates.ts so that the luca init
command can import them. The scripts/ file becomes a thin re-export shim for
backward compatibility with build-deploy.ts.
