+++ .planning/STATE.md
@@ -57,4 +57,4 @@ _State updated: 2026-03-14 — v4.4.0 milestone archived, state reset for next m
 
 ---
 
-_State generated from machine snapshot at 2026-03-17T17:21:16.521Z_
+_State generated from machine snapshot at 2026-03-17T17:31:48.615Z_
